Create main.py
A password generator. It asks the user how many letters, symbols, and numbers they want in their password. It then randomly selects the required numbers of letters, symbols, and numbers from arrays containing all the options and stores the chosen values as a new array - "password_list". It then shuffles the "password_list" array and then stores the shuffled values as a string and displays the result to the user as a randomly generated password.