Create main.py
A Hangman game where the goal is for the user to guess the secret word before the stick figure is hung. The program picks a random word from a word list. The user is asked to guess a letter and is told whether or not their guess is correct. If the user guesses the word before they run out of lives, the win. If not, they lose. The correct word is displayed to the user.