Create main.py
A caesar cipher program that encodes and decodes messages by replacing each letter of a given text with a letter at a fixed number of positions down the alphabet. The program asks the user if they would like to encode or decode a message, what their message is and what shift number they would like to use. The program displays the result and then asks the user if they would like to go again.