Create main.py
A Secret Auction program which asks the user for their name and how much they want to bid before asking if there is another person who wants to make a bid. If the answer is 'yes', the screen gets cleared (NOTE: the clear function does not work in an IDE like PyCharm, it must instead be run in PowerShell, GitBash, Hyper, etc. for the clear function to work) and the next person is asked for their name and bid. This goes on unto the answer of 'no' is given, and then the program ends. The highest bidder is determined and displayed on screen. .