Create main.py
A calculator program that asks the user for their first number, the mathematical operation they want to use, and then their second number. It displays the result and asks the user if they want to continue working with this result or not. If they answer with 'y', the user will be asked for another mathematical operation and another number. When the user answers with 'n', the screen is cleared and the calculator "restarts".