Create main.py
The Coffee Machine program is a simple simulation of a coffee machine using OOPs concepts. It allows the user to choose a drink from a menu, and then checks to see if the machine has enough resources to make the drink. If the machine has enough resources, the drink is made and the user is given a message indicating that their drink is ready. Otherwise, the user is given a message indicating that there are not enough resources to make the drink.