Update main.py
Pong is a two-dimensional sports game that simulates table tennis. Players control an in-game paddle by moving it vertically across the left or right side of the screen. They can compete against another player controlling a second paddle on the opposing side. Players use the paddles to hit a ball back and forth. The goal is for each player to try to earn more points than their opponent; points are earned when one fails to return the ball to the other.