Create main.py
This code is a simple game. The game uses Turtle graphics to display a map of the United States. The player is prompted to enter the name of a state. If the player enters a valid state name, the state name is written on the map. The game ends when the player has guessed all 50 states or when the player enters "Exit".